+++ src/main/java/site/gachontable/domain/waiting/exception/PubClosedForWaitingException.java
@@ -1,0 +1,10 @@
+package site.gachontable.domain.waiting.exception;
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class PubClosedForWaitingException extends ServiceException {
+    public PubClosedForWaitingException() {
+        super(ErrorCode.PUB_CLOSED_FOR_WAITING);
+    }
+}
