+++ src/main/java/site/gachontable/domain/waiting/exception/WaitingNotFoundException.java
@@ -1,0 +1,10 @@
+package site.gachontable.domain.waiting.exception;
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class WaitingNotFoundException extends ServiceException {
+    public WaitingNotFoundException() {
+        super(ErrorCode.WAITING_NOT_FOUND);
+    }
+}
