+++ src/main/java/site/gachontable/domain/waiting/service/CreateWaitingImpl.java
@@ -1,27 +1,28 @@
-package site.gachontable.gachontablebe.domain.waiting.usecase;
+package site.gachontable.domain.waiting.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-import site.gachontable.gachontablebe.domain.pub.domain.Pub;
-import site.gachontable.gachontablebe.domain.pub.domain.repository.PubRepository;
-import site.gachontable.gachontablebe.domain.pub.exception.PubNotFoundException;
-import site.gachontable.gachontablebe.domain.shared.event.SentBiztalkEvent;
-import site.gachontable.gachontablebe.domain.user.domain.User;
-import site.gachontable.gachontablebe.domain.user.domain.repository.UserRepository;
-import site.gachontable.gachontablebe.domain.user.exception.UserNotFoundException;
-import site.gachontable.gachontablebe.domain.waiting.domain.Waiting;
-import site.gachontable.gachontablebe.domain.waiting.domain.repository.WaitingRepository;
-import site.gachontable.gachontablebe.domain.waiting.exception.UserWaitingLimitExcessException;
-import site.gachontable.gachontablebe.domain.waiting.exception.WaitingAlreadyExistsException;
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.request.RemoteWaitingRequest;
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.response.WaitingResponse;
-import site.gachontable.gachontablebe.domain.waiting.type.Position;
-import site.gachontable.gachontablebe.domain.waiting.type.Status;
-import site.gachontable.gachontablebe.global.config.redis.RedissonLock;
-import site.gachontable.gachontablebe.global.success.SuccessCode;
+import site.gachontable.domain.waiting.port.in.CreateWaiting;
+import site.gachontable.infra.security.principal.AuthDetails;
+import site.gachontable.domain.pub.domain.Pub;
+import site.gachontable.domain.pub.domain.repository.PubRepository;
+import site.gachontable.domain.pub.exception.PubNotFoundException;
+import site.gachontable.domain.shared.event.SentBiztalkEvent;
+import site.gachontable.domain.member.domain.User;
+import site.gachontable.domain.member.port.out.UserRepository;
+import site.gachontable.domain.member.exception.UserNotFoundException;
+import site.gachontable.domain.waiting.domain.Waiting;
+import site.gachontable.domain.waiting.port.out.WaitingRepository;
+import site.gachontable.domain.waiting.exception.UserWaitingLimitExcessException;
+import site.gachontable.domain.waiting.exception.WaitingAlreadyExistsException;
+import site.gachontable.presentation.waiting.dto.request.RemoteWaitingRequest;
+import site.gachontable.presentation.waiting.dto.response.WaitingResponse;
+import site.gachontable.domain.waiting.type.Position;
+import site.gachontable.domain.waiting.type.Status;
+import site.gachontable.infra.redis.RedissonLock;
+import site.gachontable.independent.type.SuccessCode;
 
 import java.util.Arrays;
 import java.util.HashMap;
