--- src/main/java/site/gachontable/gachontablebe/domain/waiting/exception/UserWaitingLimitExcessException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class UserWaitingLimitExcessException extends ServiceException {
-    public UserWaitingLimitExcessException() {
-        super(ErrorCode.WAITING_OVER_COUNT);
-    }
-}
