--- src/main/java/site/gachontable/gachontablebe/domain/waiting/exception/WaitingNotFoundException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class WaitingNotFoundException extends ServiceException {
-    public WaitingNotFoundException() {
-        super(ErrorCode.WAITING_NOT_FOUND);
-    }
-}
