--- src/main/java/site/gachontable/gachontablebe/domain/waiting/presentation/dto/request/CancelRequest.java
@@ -1,6 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.presentation.dto.request;
-
-import java.util.UUID;
-
-public record CancelRequest(UUID waitingId) {
-}
