--- src/main/java/site/gachontable/gachontablebe/domain/waiting/usecase/CancelWaiting.java
@@ -1,9 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.usecase;
-
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.request.CancelRequest;
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.response.WaitingResponse;
-
-public interface CancelWaiting {
-
-    WaitingResponse execute(CancelRequest request, String lockKey);
-}
