--- src/main/java/site/gachontable/gachontablebe/domain/waiting/usecase/CreateWaiting.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.usecase;
-
-import site.gachontable.gachontablebe.domain.auth.domain.AuthDetails;
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.request.RemoteWaitingRequest;
-import site.gachontable.gachontablebe.domain.waiting.presentation.dto.response.WaitingResponse;
-
-public interface CreateWaiting {
-
-    WaitingResponse execute(AuthDetails authDetails, RemoteWaitingRequest request, String lockKey);
-}
