+++ src/main/java/site/gachontable/presentation/waiting/api/WaitingController.java
@@ -9,15 +9,19 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import site.gachontable.domain.auth.domain.AuthDetails;
+import site.gachontable.domain.waiting.port.in.CancelWaiting;
+import site.gachontable.domain.waiting.port.in.CreateWaiting;
+import site.gachontable.domain.waiting.port.in.GetStatus;
+import site.gachontable.domain.waiting.port.in.GetStatusByBiztalk;
+import site.gachontable.infra.security.principal.AuthDetails;
 import site.gachontable.presentation.waiting.dto.request.CancelRequest;
 import site.gachontable.presentation.waiting.dto.request.RemoteWaitingRequest;
 import site.gachontable.presentation.waiting.dto.response.StatusResponse;
 import site.gachontable.presentation.waiting.dto.response.WaitingHistoryResponse;
 import site.gachontable.presentation.waiting.dto.response.WaitingResponse;
 import site.gachontable.domain.waiting.type.Position;
-import site.gachontable.domain.waiting.usecase.*;
-import site.gachontable.global.error.ErrorResponse;
+import site.gachontable.domain.waiting.service.*;
+import site.gachontable.independent.error.ErrorResponse;
 
 import java.util.List;
 import java.util.UUID;
